feat(ui): manage the Slack channel binding via a modal
Replaces the enable-toggle + inline form + save flow: an unbound card
offers Connect channel, a bound one shows the binding as a row with an
Edit / Disconnect menu (confirmed, destructive), and the form lives in
a connect/edit modal.
