+++ packages/ui/src/modules/sandboxes/components/sandbox-channels-section.tsx
@@ -9,14 +9,6 @@ export function SandboxChannelsSection({ agentId }: { agentId: string }) {
   const { data } = useAgents();
   const agent = data?.list.find((a) => a.id === agentId);
   const available = data?.availableChannels ?? {};
-  // Remount the Slack form when the binding changes out-of-band (in-chat
-  // ambient command, another tab) so its local state reseeds.
-  const slackChannel = agent?.channels.find((c) => c.type === "slack");
-  const slackKey = [
-    agent?.id ?? "none",
-    slackChannel?.type === "slack" ? slackChannel.slackChannelId : "",
-    slackChannel?.type === "slack" ? (slackChannel.ambient ?? false) : "",
-  ].join(":");
 
   return (
     <section className="mb-8">
@@ -31,7 +23,7 @@ export function SandboxChannelsSection({ agentId }: { agentId: string }) {
         </p>
       ) : (
         <Inset className="flex flex-col gap-4">
-          {available.slack && <SlackChannelCard key={slackKey} agent={agent} />}
+          {available.slack && <SlackChannelCard agent={agent} />}
           {available.telegram && <TelegramChannelCard agentId={agentId} />}
         </Inset>
       )}
